+++ spec/dummy/spec/models/post_spec.rb
@@ -21,6 +21,38 @@ RSpec.describe Post do
     end
   end if defined? Turbo::Broadcastable
 
+  describe 'relations' do
+    context 'when decorated' do
+      subject { relation.decorate }
+
+      let(:relation)   { described_class.where('1=1') }
+      let(:persisted)  { relation.create! [{}] * rand(0..2) }
+
+      before { persisted } # should exist
+
+      it 'returns a decorated collection' do
+        is_expected.to match_array persisted
+        is_expected.to be_all &:decorated?
+      end
+
+      it 'uses cached records' do
+        expect(relation).not_to be_loaded
+
+        relation.load
+
+        expect { subject.to_a }.to execute.exactly(0).queries
+      end
+
+      it 'caches records' do
+        expect(relation).not_to be_loaded
+
+        relation.decorate.to_a
+
+        expect { subject.to_a; relation.load }.to execute.exactly(0).queries
+      end
+    end
+  end
+
   describe 'associations' do
     context 'when decorated' do
       subject { associated.decorate }
@@ -47,7 +79,7 @@ RSpec.describe Post do
       it 'caches records' do
         expect(associated).not_to be_loaded
 
-        associated.decorate
+        associated.decorate.to_a
 
         expect { subject.to_a; associated.load }.to execute.exactly(0).queries
       end